pyproject.toml
[tool.poetry]
name = "sevdesk"
version = "0.1.0"
description = "A client to integrate the SevDesk API in Python Projects."
authors = ["Heinz-Peter Liechtenecker <heinz.liechtenecker@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
openapi-python-client = "^0.11.1"
attrs = "^21.4.0"
cattrs = "^1.10.0"
httpx = "^0.22.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
